agent/pyproject.toml
[project]
name = "minih-livekit-agent"
version = "0.1.0"
description = "AI 实时语音交互系统 - Python Agent"
requires-python = ">=3.12"
dependencies = [
    "livekit-api>=0.7.0",
    "sherpa-onnx>=1.10.0",
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "numpy",
    "python-dotenv>=1.2.1",
    "soxr>=1.0.0",
    "livekit-agents>=1.3.10",
    "livekit-plugins-silero>=1.3.10",
]




[project.scripts]
minih-agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
